refactor: project files reorganization
