+++ src/index.ts
@@ -1,4 +1,4 @@
-import { Store } from './store';
+import { Store } from './models/store';
 
 const store = new Store();
