+++ src/models/priceCodes/newPriceCode.ts
@@ -1,0 +1,16 @@
+import { PriceCode } from "./priceCode";
+
+export class NewReleasePriceCode extends PriceCode {
+
+    get name(): string {
+        return 'NEW RELEASE';
+    }
+
+    getAmount(daysRented: number): number {
+        return daysRented * 3;
+    }
+
+    getFrequentRenterPoints(daysRented: number): number {
+        return daysRented > 1 ? 2 : 1;
+    }
+}
